last one inshalla

+++ src/MP3Track.cpp
@@ -37,7 +37,7 @@ void MP3Track::analyze_beatgrid() {
 double beats = (duration_seconds / 60.0) * bpm;
 double precision_factor = bitrate / 320.0;
 
-std::cout << "  → Estimated beats: " << beats
+std::cout << "  → Estimated beats: " << (int)beats
           << "  → Compression precision factor: " << precision_factor
           << std::endl;
