+++ src/WAVTrack.cpp
@@ -32,7 +32,7 @@ void WAVTrack::analyze_beatgrid() {
     // should print "  → Estimated beats: <beats>  → Precision factor: 1.0 (uncompressed audio)"
        double beats = (duration_seconds / 60.0) * bpm;
 
-      std::cout << "  → Estimated beats: " << beats<< "  → Precision factor: 1 (uncompressed audio)"<< std::endl;
+      std::cout << "  → Estimated beats: " << (int)beats<< "  → Precision factor: 1 (uncompressed audio)"<< std::endl;
           
           
 }
